Don't always require DatabaseMapping as context manager, drop locks
Re #121
